Update 'Current.dart'
Created a column that consists of an 'Expanded' widget that defines the number of cards present within the page. An 'ItemWidget' class has also been created that consists of two parameters called 'labelText' and 'data'. It returns a container that consists of the structure of the cards present within the 'Current.dart' page. When the drop down icon is clicked, an alert dialog will be displayed that shows the required data to the user.