Update 'Login.dart'
Updated validation logic in the login page and created three text editing controllers to keep a track of the data typed into the text fields by the user. A suffix-icon has also been added in-order to toggle between the visibility of the password. Finally, a text widget has been added that let's the user create a new account if the individual is a new user.