Update 'Register.dart'
Updated validation logic in the register page and created four text editing controllers to keep a track of the data typed into the text fields by the user. Two suffix-icons has also been added in-order to toggle between the visibility of the password fields. Finally, a text widget has been added that let's the user login to the account if the individual has an account.